fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,13 +12,13 @@
 int main(void)
 {
 	int i;
-	unsigned int fb1 = 0, fb2 = 1;
-	unsigned int sum;
+	long unsigned int fb1 = 0, fb2 = 1;
+	long unsigned int sum;
 
 	for (i = 0; i < 50; i++)
 	{
 	sum = fb1 + fb2;
-	printf("%u", sum);
+	printf("%lu", sum);
 	fb1 = fb2;
 	fb2 = sum;
 	if (i != 49)
